+++ app/models/housekeeping.rb
@@ -55,10 +55,22 @@ class Housekeeping
           WHERE  Partition_Position != (SELECT MAX(pi.Partition_Position) FROM Partitions pi WHERE pi.Interval = p.Interval)
           ORDER BY Partition_Position
         "
+
+        # check for locks on partitions only once to ensure that expensive SQL is executed as little as possible
+        locked_partitions = Database.select_all("\
+          SELECT DISTINCT o.SubObject_Name Partition_Name
+          FROM   gv$Lock l
+          JOIN   User_Objects o ON o.Object_ID = l.ID1
+          WHERE  o.Object_Name    = 'EVENT_LOGS'
+        ").map{|r| r.partition_name}
         partitions_to_check.each do |part|
           # if only range partitions exists (no interval) than preserve the youngest two range partitions (because the first partition is not scanned by worker)
           if part.interval_count > 0 || ( part.partition_position < part.max_partition_position - 2 )
-            EventLog.check_and_drop_partition(part['partition_name'], 'Housekeeping.do_housekeeping_internal')
+            if locked_partitions.include? part.partition_name                   # Don't check partition that has pending transactions
+              Rails.logger.info "#{caller}: Check partition #{part.partition_name} for drop not possible because there are pending transactions"
+            else
+              EventLog.check_and_drop_partition(part.partition_name, 'Housekeeping.do_housekeeping_internal', lock_already_checked: true)
+            end
           end
         end
       end